api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
httpx
pydantic
